ssip-chatbot/requirements.txt
flask==2.3.3
openai==0.28.0
python-dotenv==1.0.0
flask-sqlalchemy==3.0.5
flask-login==0.6.2
flask-session==0.5.0
pyttsx3==2.90
SpeechRecognition==3.10.0
gunicorn==21.2.0
python-socketio==5.8.0
